moved most of AddAppForm into new AppForm component
